Refactors web pages for client-only rendering; updates Yjs deps
Reworks all web pages to use dynamic client-only imports, improving hydration consistency and reducing SSR errors from collaborative and AI components.

Updates Yjs and y-websocket dependencies to unify versions across packages, resolving constructor mismatch and "already imported" warnings.

Refines Next.js config to enforce single instances of Yjs-related modules via robust aliasing.

Improves pre-commit hook logic to favor local lint-staged execution, addressing ESM import issues and ensuring more reliable linting.

Relates to stability and collaboration fixes for web platform.
