+++ apps/mobile/package.json
@@ -20,8 +20,8 @@
     "expo-linking": "~6.3.1",
     "react": "18.2.0",
     "react-native": "0.74.0",
-    "y-websocket": "^2.0.4",
-    "yjs": "^13.6.18"
+    "y-websocket": "^1.5.4",
+    "yjs": "^13.6.27"
   },
   "devDependencies": {
     "@types/node": "^20.0.0",
