+++ apps/web/package.json
@@ -37,7 +37,7 @@
     "react-dom": "^18.3.1",
     "uuid": "^9.0.0",
     "y-protocols": "^1.0.6",
-    "y-websocket": "^1.5.0",
+    "y-websocket": "^1.5.4",
     "yjs": "^13.6.27"
   },
   "devDependencies": {
