+++ apps/web/pages/offline-demo.tsx
@@ -1,185 +1,10 @@
+'use client';
+
+import dynamic from 'next/dynamic';
 import React from 'react';
 
-// Mock OfflineEditor component for now
-interface OfflineEditorProps {
-  room: string;
-  serverUrl: string;
-  children: (
-    doc: unknown,
-    status: { isConnected: boolean; pendingChanges: number; lastSync?: Date }
-  ) => React.ReactNode;
+const Client = dynamic(() => import('./offline-demo.client'), { ssr: false });
+
+export default function OfflineDemo() {
+  return <Client />;
 }
-
-const OfflineEditor: React.FC<OfflineEditorProps> = props => {
-  const { children } = props;
-  // Mock document and status
-  const mockDoc = {};
-  const mockStatus = {
-    isConnected: false,
-    pendingChanges: 0,
-    lastSync: new Date(),
-  };
-
-  return <>{children(mockDoc, mockStatus)}</>;
-};
-
-/**
- * Complete demonstration of offline-enabled collaborative editing
- */
-const OfflineCollaborationDemo: React.FC = () => {
-  return (
-    <div style={{ padding: '20px', fontFamily: 'system-ui, sans-serif' }}>
-      <h1>🔗 Offline-Enabled Collaborative Platform</h1>
-
-      <div style={{ marginBottom: '20px' }}>
-        <h2>✨ Key Features Implemented:</h2>
-        <ul style={{ lineHeight: '1.6' }}>
-          <li>
-            <strong>🤖 Complete AI Integration Suite</strong>
-            <ul>
-              <li>Local Model Support (Ollama) for offline AI capabilities</li>
-              <li>
-                Intelligent Code Completion with caching and context awareness
-              </li>
-              <li>13+ Refactoring Tools with safety analysis</li>
-              <li>Context-Aware Assistant with codebase understanding</li>
-            </ul>
-          </li>
-          <li>
-            <strong>📱 Offline Persistence with IndexedDB</strong>
-            <ul>
-              <li>Automatic synchronization when connection returns</li>
-              <li>Real-time connection monitoring and status indication</li>
-              <li>Seamless offline editing with change tracking</li>
-              <li>Manual sync capabilities and error handling</li>
-            </ul>
-          </li>
-          <li>
-            <strong>🔄 Real-time Collaboration</strong>
-            <ul>
-              <li>WebSocket-based live collaboration with Yjs</li>
-              <li>User awareness and presence indicators</li>
-              <li>Conflict-free document merging</li>
-            </ul>
-          </li>
-        </ul>
-      </div>
-
-      <div
-        style={{
-          border: '2px solid #e5e7eb',
-          borderRadius: '8px',
-          overflow: 'hidden',
-          backgroundColor: '#f9fafb',
-        }}
-      >
-        <div
-          style={{
-            padding: '12px 16px',
-            backgroundColor: '#3b82f6',
-            color: 'white',
-            fontWeight: '600',
-          }}
-        >
-          📝 Live Collaborative Editor with Offline Support
-        </div>
-
-        <OfflineEditor room='platform-demo' serverUrl='ws://localhost:1234'>
-          {(doc, status) => (
-            <div style={{ padding: '16px' }}>
-              <div
-                style={{
-                  marginBottom: '16px',
-                  padding: '12px',
-                  backgroundColor: status.isConnected ? '#ecfdf5' : '#fef3c7',
-                  borderRadius: '6px',
-                  border: `1px solid ${
-                    status.isConnected ? '#d1fae5' : '#fde68a'
-                  }`,
-                }}
-              >
-                <h3 style={{ margin: '0 0 8px 0', fontSize: '16px' }}>
-                  {status.isConnected ? '🌐 Online Mode' : '📱 Offline Mode'}
-                </h3>
-                <p style={{ margin: 0, fontSize: '14px', color: '#374151' }}>
-                  {status.isConnected
-                    ? 'Real-time collaboration active. Changes sync instantly with other users.'
-                    : 'Working offline. Your changes are being saved locally and will sync when you reconnect.'}
-                </p>
-                {status.pendingChanges > 0 && (
-                  <p
-                    style={{
-                      margin: '8px 0 0 0',
-                      fontSize: '12px',
-                      color: '#f59e0b',
-                      fontWeight: '500',
-                    }}
-                  >
-                    📋 {status.pendingChanges} changes pending sync
-                  </p>
-                )}
-              </div>
-
-              <div
-                style={{
-                  height: '300px',
-                  backgroundColor: 'white',
-                  border: '1px solid #d1d5db',
-                  borderRadius: '6px',
-                  padding: '16px',
-                  display: 'flex',
-                  flexDirection: 'column',
-                  alignItems: 'center',
-                  justifyContent: 'center',
-                  textAlign: 'center',
-                }}
-              >
-                <div style={{ fontSize: '48px', marginBottom: '16px' }}>
-                  {status.isConnected ? '⚡' : '💾'}
-                </div>
-                <h3 style={{ margin: '0 0 8px 0', color: '#374151' }}>
-                  {status.isConnected
-                    ? 'Connected & Synced'
-                    : 'Offline Mode Active'}
-                </h3>
-                <p style={{ margin: 0, color: '#6b7280', maxWidth: '300px' }}>
-                  {status.isConnected
-                    ? 'Your document is live and syncing with collaborators in real-time.'
-                    : 'Continue editing! All changes are saved locally and will automatically sync when you reconnect.'}
-                </p>
-
-                {status.lastSync && (
-                  <p
-                    style={{
-                      margin: '12px 0 0 0',
-                      fontSize: '12px',
-                      color: '#9ca3af',
-                    }}
-                  >
-                    Last sync: {status.lastSync.toLocaleString()}
-                  </p>
-                )}
-              </div>
-            </div>
-          )}
-        </OfflineEditor>
-      </div>
-
-      <div style={{ marginTop: '24px', fontSize: '14px', color: '#6b7280' }}>
-        <h3 style={{ margin: '0 0 8px 0', color: '#374151' }}>
-          🚀 Platform Status:
-        </h3>
-        <ul style={{ margin: 0 }}>
-          <li>✅ AI Integration Complete (OpenAI, Anthropic, Ollama)</li>
-          <li>✅ Code Completion & Refactoring Tools</li>
-          <li>✅ Context-Aware AI Assistant</li>
-          <li>✅ Real-time Collaboration (Yjs + WebSocket)</li>
-          <li>✅ Offline Persistence (IndexedDB)</li>
-          <li>✅ Cross-Platform Support (Web, Mobile, Desktop)</li>
-        </ul>
-      </div>
-    </div>
-  );
-};
-
-export default OfflineCollaborationDemo;
