+++ package.json
@@ -47,9 +47,8 @@
     "turbo": "^1.12.5",
     "typescript": "^5.3.3",
     "dotenv": "^16.1.4"
-  }
-
-  ,"lint-staged": {
+  },
+  "lint-staged": {
     "*.{js,jsx,ts,tsx,json,md}": [
       "prettier --write",
       "eslint --fix"
