+++ packages/editor-core/package.json
@@ -12,7 +12,9 @@
   "dependencies": {
     "y-indexeddb": "^9.0.12",
     "y-protocols": "^1.0.6",
-    "y-websocket": "^1.5.0",
+    "y-websocket": "^1.5.4"
+  },
+  "peerDependencies": {
     "yjs": "^13.6.27"
   },
   "devDependencies": {
